+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
 ### July 29, 2026
-- moved `account.go` to `potato-serialization-lib`
-- reduced import complexity
+- Moved `account.go` to `potato-serialization-lib`
+- Reduced import complexity
+
+### July 30, 2026
+- Started working on `Ledger` Syncing